pyproject.toml
[tool.poetry]
name = "self-corrective-rag-llamaindex"
version = "0.1.0"
description = ""
authors = ["kamran945 <kamran945@gmail.com>"]
readme = "README.md"
packages = [
    { include = "src" },
]

[tool.poetry.dependencies]
python = ">=3.11,<=3.13"
llama-index = "^0.12.15"
llama-index-indices-managed-llama-cloud = "^0.6.4"
llama-cloud = "^0.1.11"
llama-parse = "^0.5.20"
llama-index-llms-groq = "^0.3.1"
torch = "^2.6.0"
llama-index-embeddings-huggingface = "^0.5.1"
llama-index-utils-workflow = "^0.3.0"
langchain-core = "^0.3.33"
langchain-groq = "^0.2.4"
python-dotenv = "^1.0.1"
langchain = "^0.3.17"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
